fix bot detector for tests with badly behaved HttpClientFactories
