+++ packages/flutter_tools/test/commands.shard/permeable/packages_test.dart
@@ -21,22 +21,6 @@ import '../../src/context.dart';
 import '../../src/mocks.dart' show MockProcessManager, MockStdio, PromptingProcess;
 import '../../src/testbed.dart';
 
-class AlwaysTrueBotDetector implements BotDetector {
-  const AlwaysTrueBotDetector();
-
-  @override
-  Future<bool> get isRunningOnBot async => true;
-}
-
-
-class AlwaysFalseBotDetector implements BotDetector {
-  const AlwaysFalseBotDetector();
-
-  @override
-  Future<bool> get isRunningOnBot async => false;
-}
-
-
 void main() {
   Cache.disableLocking();
   group('packages get/upgrade', () {
