+++ packages/flutter_tools/test/src/context.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:io' as io;
 
 import 'package:flutter_tools/src/android/android_workflow.dart';
+import 'package:flutter_tools/src/base/bot_detector.dart';
 import 'package:flutter_tools/src/base/config.dart';
 import 'package:flutter_tools/src/base/context.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
@@ -37,6 +38,7 @@ import 'package:mockito/mockito.dart';
 
 import 'common.dart';
 import 'fake_process_manager.dart';
+import 'mocks.dart';
 import 'throwing_pub.dart';
 
 export 'package:flutter_tools/src/base/context.dart' show Generator;
@@ -163,6 +165,15 @@ void testUsingContext(
           });
         },
       );
+    }, overrides: <Type, Generator>{
+      // This has to go here so that runInContext will pick it up when it tries
+      // to do bot detection before running the closure.  This is important
+      // because the test may be giving us a fake HttpClientFactory, which may
+      // throw in unexpected/abnormal ways.
+      // If a test needs a BotDetector that does not always return true, it
+      // can provide the AlwaysFalseBotDetector in the overrides, or its own
+      // BotDetector implementation in the overrides.
+      BotDetector: overrides[BotDetector] ?? () => const AlwaysTrueBotDetector(),
     });
   }, testOn: testOn, skip: skip);
 }
